+++ threedigrid_builder/__init__.py
@@ -2,5 +2,5 @@ from .application import *  # NOQA
 from .exceptions import *  # NOQA
 
 # fmt: off
-__version__ = '1.6.2.dev0'
+__version__ = '1.7.0'
 # fmt: on