+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog of threedigrid-builder
 ================================
 
-1.6.2 (unreleased)
+1.7.0 (2023-01-11)
 ------------------
 
 - Fix build with numpy >= 1.24